+++ CHANGELOG.md
@@ -5,3 +5,31 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Package documentation;
+- `start` and `runningAs` methods were added to `CreateDaemonCommand` class;
+- `identifiedAs` and `usingPort` methods were added to `CreateFirewallRuleCommand` class.
+
+### Changed
+- `DaemonsManager::create` method now accepts new daemon name instead of full daemon payload. `CreateDaemonCommand::runningAs` method allows you to set daemon user;
+- `FirewallManager::create` method now accepts rule name instead of full firewall rule payload. `CreateFirewallRuleCommand::usingPort` method allows you to set rule port number;
+- `CreateJobCommand::runAs` method was renamed to `runningAs`, method signature remains the same.
+
+## [0.9.0] - 2017-03-12
+Initial release.
+
+### Added
+- Servers Management;
+- Servers Factory;
+- Services Management;
+- Deamons;
+- Firewall Rules;
+- Scheduled Jobs;
+- MySQL Databases Management;
+- MySQL Users Management;
+- Sites Management;
+- Site Applications;
+- Deployment Management.
+
+[Unreleased]: https://github.com/tzurbaev/laravel-forge-api/compare/0.9.0...HEAD
+[0.9.0]: https://github.com/tzurbaev/laravel-forge-api/releases/tag/0.9.0